requirements.txt
fastapi 
uvicorn 
weaviate-client 
requests 
transformers 
psycopg2-binary 
scikit-learn
